refactor: Chuyển CSDL từ MySQL sang SQL Server

+++ BUS/Services/IFoodItemService.cs
@@ -10,11 +10,11 @@ namespace BUS.Services
     public interface IFoodItemService
     {
         Task<IEnumerable<FoodItemDTO>> GetAllAsync();
-        Task<IEnumerable<FoodItemDTO>> GetFoodsByCategoryAsync(long categoryId);
-        Task<FoodItemDTO?> GetByIdAsync(long id);
+        Task<IEnumerable<FoodItemDTO>> GetFoodsByCategoryAsync(int categoryId);
+        Task<FoodItemDTO?> GetByIdAsync(int id);
         Task AddAsync(FoodItemDTO dto);
         Task UpdateAsync(FoodItemDTO dto);
-        Task<bool> DeleteAsync(long id);
+        Task<bool> DeleteAsync(int id);
     }
 
 }
