+++ BUS/Services/IOrderService.cs
@@ -10,9 +10,10 @@ namespace BUS.Services
     public interface IOrderService
     {
         Task<IEnumerable<OrderDTO>> GetAllAsync();
-        Task<OrderDTO?> GetByIdAsync(long id);
-        Task<long> CreateAsync(OrderDTO dto);
-        Task UpdateStatusAsync(long orderId, string status);
-        Task<bool> DeleteAsync(long id);
+        Task<OrderDTO?> GetByIdAsync(int id);
+        Task<int> CreateAsync(OrderDTO dto);
+
+        Task UpdateStatusAsync(int orderId, string status);
+        Task<bool> DeleteAsync(int id);
     }
 }
