+++ WebAPI/Controllers/OrderController.cs
@@ -19,7 +19,7 @@ namespace WebAPI.Controllers
         public async Task<IActionResult> GetAll() => Ok(await _service.GetAllAsync());
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetById(long id)
+        public async Task<IActionResult> GetById(int id)
         {
             var order = await _service.GetByIdAsync(id);
             if (order == null) return NotFound();
@@ -34,14 +34,14 @@ namespace WebAPI.Controllers
         }
 
         [HttpPut("{id}/status")]
-        public async Task<IActionResult> UpdateStatus(long id, [FromBody] string status)
+        public async Task<IActionResult> UpdateStatus(int id, [FromBody] string status)
         {
             await _service.UpdateStatusAsync(id, status);
             return Ok(new { message = "Order status updated" });
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(long id)
+        public async Task<IActionResult> Delete(int id)
         {
             var result = await _service.DeleteAsync(id);
             if (!result) return NotFound();
